added renaming files, folders, modifieds deleting files and folders and respective testings

+++ blueprints/main.py
@@ -1,12 +1,13 @@
 import os
+import shutil
 
 from werkzeug.utils import secure_filename
 
 from flask import Blueprint, render_template, redirect, request, abort, flash, url_for, send_from_directory
 from flask_login.utils import login_required, current_user
 
-from ..utils import get_path_folders_and_files, is_own, get_user_path, protected_path
-from ..forms import CreateDirForm, FileUploadForm, DeleteFileForm
+from ..utils import get_path_files_and_folders, is_own, get_user_path, protected_path
+from ..forms import CreateDirForm, FileUploadForm, DeleteFileForm, RenameForm
 from ..models import User
 from .. import db
 
@@ -27,7 +28,7 @@ def cloud_private(path, encpath):
     dpath = os.path.join(current_user.get_private_user_path(), path)
     
     try:
-        files, folders = get_path_folders_and_files(dpath)
+        files, folders = get_path_files_and_folders(dpath)
     except FileNotFoundError:
         abort(404, description='Carpeta no encontrada!')
     
@@ -55,7 +56,7 @@ def cloud_public(username, path, encpath):
     dpath = os.path.join(user.get_public_user_path(), path)
     
     try:
-        files, folders = get_path_folders_and_files(dpath)
+        files, folders = get_path_files_and_folders(dpath)
     except FileNotFoundError:
         abort(404, description='Carpeta no conseguida!')
     
@@ -81,16 +82,16 @@ def cloud_create_dir(status, username, path, encpath):
     form = CreateDirForm()
     next = request.form.get('next', None)
     
+    user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
+    is_own(user)
+    
+    if not status in ['private', 'public']:
+        abort(404)
+    
     if next is None:
         abort(400)
         
     if form.validate_on_submit():
-        user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
-        is_own(user)
-        
-        if not status in ['private', 'public']:
-            abort(404)
-        
         dpath = get_user_path(user, status, path)
         new_dir_path = os.path.join(dpath, form.dir.data, '')
         
@@ -99,7 +100,7 @@ def cloud_create_dir(status, username, path, encpath):
             flash('El directorio se creo con exito')
         except FileExistsError:
             flash('Este archivo ya existe', 'error')
-        
+            
     else:
         flash('Invalido!', 'error')
         
@@ -113,16 +114,16 @@ def cloud_upload_file(status, username, path, encpath):
     form = FileUploadForm()
     next = request.form.get('next', None)
     
+    user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
+    is_own(user)
+        
+    if not status in ['private', 'public']:
+        abort(404)
+            
     if next is None:
         abort(400)
     
     if form.validate_on_submit():
-        user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
-        is_own(user)
-        
-        if not status in ['private', 'public']:
-            abort(404)
-        
         dpath = get_user_path(user, status, path)
         
         file = form.file.data
@@ -142,33 +143,40 @@ def cloud_upload_file(status, username, path, encpath):
 @main_bp.route('/cloud/delete_file/<status>/<username>/path/<path:path>', methods=['GET', 'POST'])
 @login_required
 @protected_path
-def cloud_delete_file(status, username, path, encpath):
+def cloud_delete_fl_fld(status, username, path, encpath):
     form = DeleteFileForm()
+    next = request.args.get('next', None)
+    
+    user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
+    is_own(user)
+        
+    if not status in ['private', 'public']:
+        abort(404)
+            
+    if next is None:
+        next = url_for('main_bp.cloud_private') if status == 'private' else url_for('main_bp.cloud_public', username=user.username)
     
     if form.validate_on_submit():
-        user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
-        is_own(user)
-        
-        if not status in ['private', 'public']:
-            abort(404)
         
         fpath = get_user_path(user, status, path)
         
-        redirect_url = url_for('main_bp.cloud_private') if status == 'private' else url_for('main_bp.cloud_public', username=user.username)
+        if os.path.isfile(fpath):
+            try:
+                os.remove(fpath)
+                flash('Se ha eliminado con exito el archivo!')
+            except FileNotFoundError:
+                flash('Este archivo no existe!', 'error')
         
-        try:
-            os.remove(fpath)
-            flash('Se ha eliminado con exito el archivo!')
-        except FileNotFoundError:
-            flash('Este archivo no existe!', 'error')
-            
-        return redirect(redirect_url)
-    
-    context = {
-        'form': form
-    }
-    
-    return render_template('delete_file.html', **context)
+        else:
+            try:
+                shutil.rmtree(fpath)
+                flash('Se ha eliminado con exito la carpeta!')
+            except FileNotFoundError:
+                flash('Esta carpeta no existe!', 'error')
+                
+        return redirect(next)
+        
+    return render_template('delete_file.html', form=form)
 
 @main_bp.route('/cloud/download_file/<status>/<username>/path/', defaults={'path': ''})
 @main_bp.route('/cloud/download_file/<status>/<username>/path/<path:path>')
@@ -180,13 +188,46 @@ def cloud_download_file(status, username, path, encpath):
     if not status in ['private', 'public']:
          abort(404)
     
-    dpath = get_user_path(user, status, path)
-    fpath, filename = dpath.rsplit('/', maxsplit=1)
-    
     if status == 'private' and current_user.email != user.email:
         abort(404, description='Este archivo no existe')
+        
+    dpath = get_user_path(user, status, path)
+    
+    filename = os.path.basename(dpath)
+    fpath = os.path.dirname(dpath)
     
     if not os.path.exists(fpath):
         abort(404, description='Este archivo no existe')
     
     return send_from_directory(fpath, filename, as_attachment=True)
+
+@main_bp.route('/cloud/rename/<status>/<username>/path/', defaults={'path': ''}, methods=['GET', 'POST'])
+@main_bp.route('/cloud/rename/<status>/<username>/path/<path:path>', methods=['GET', 'POST'])
+@login_required
+@protected_path
+def cloud_rename_fl_fld(status, username, path, encpath):
+    form = RenameForm()
+    next = request.args.get('next', None)
+    
+    user = db.one_or_404(db.select(User).filter_by(username=username), description='User not found!')
+    is_own(user)
+    
+    if next is None:
+        next = url_for('main_bp.cloud_private') if status == 'private' else url_for('main_bp.cloud_public', username=user.username)
+    
+    if not status in ['private', 'public']:
+         abort(404)
+    
+    if form.validate_on_submit():
+        user_path = get_user_path(user, status, path)
+        dir_name = os.path.dirname(user_path)
+        
+        try:
+            os.rename(user_path, os.path.join(dir_name, form.name.data))
+            flash('Se ha renombrado con exito!')
+        except FileNotFoundError:
+            flash('Este archivo no existe!', 'error')
+        
+        return redirect(next)
+        
+    return render_template('cloud_rename.html', form=form)
